chore: splitting free disk getters from validators
as to get the amount of free space is needed in more than one use case
this commit splits the validators from the free disk getters.

no functional change has been introduced other than migrating code and
tests to different files/structs.

files were also renamed to better reflect their content.
